requirements.txt
Telethon==0.18.3
PyYAML==3.12
